FIX: Reorganize tests
